extracted job calculation in its own
