+++ .env.template
@@ -105,6 +105,6 @@ SCHEDULE_MAX_DELAY="00:01:00"
 SCHEDULE_DEVICE_THRESHOLD=100
 
 # Maximum amount of ads to include in the batch
-SCHEDULE_AD_THRESHOLD=10
+CALCULATE_AD_THRESHOLD=10
 # Expiration timespan for the presigned urls of ads
-SCHEDULE_URL_EXPIRE="01:00:00"
+CALCULATE_URL_EXPIRE="01:00:00"
