+++ dotnet/backend/hangfire/ScheduleJob.cs
@@ -16,52 +16,27 @@ namespace backend.Hangfire;
 public class ScheduleJob
 {
     private readonly ILogger<ScheduleJob> _logger;
-    private readonly IEventCoordinator _eventCoordinator;
     private readonly IConfigurationRepository _configRepository;
     private readonly IBackgroundJobClientV2 _backgroundJobClient;
     private readonly IDeviceRepository _deviceRepository;
     private uint pageSize =>
         uint.Parse(Environment.GetEnvironmentVariable("SCHEDULE_DEVICE_THRESHOLD")!);
-    private uint adPageSize =>
-        uint.Parse(Environment.GetEnvironmentVariable("SCHEDULE_AD_THRESHOLD")!);
-    private TimeSpan expiration =>
-        TimeSpan.Parse(Environment.GetEnvironmentVariable("SCHEDULE_URL_EXPIRE")!);
     public TimeSpan maxSpan =>
         TimeSpan.Parse(Environment.GetEnvironmentVariable("SCHEDULE_MAX_DELAY")!);
 
     private static readonly string pageKey = "CALL_FOR_SCHEDULE_PAGE";
-    private readonly IMapper _mapper;
-    private readonly IPreSignObjectService _presigner;
-    private readonly IGetOrCreateBucketService _bucketService;
-    private readonly IPutObjectService _putObjectService;
-    private readonly IGetObjectService _getObjectService;
-    private readonly IAdRepository _adRepository;
 
     public ScheduleJob(
         ILogger<ScheduleJob> logger,
-        IEventCoordinator eventCoordinator,
         IConfigurationRepository configRepo,
         IBackgroundJobClientV2 backgroundJobClient,
-        IDeviceRepository deviceRepository,
-        IMapper mapper,
-        IPreSignObjectService preSignObjectService,
-        IGetOrCreateBucketService getOrCreateBucketService,
-        IPutObjectService putObjectService,
-        IGetObjectService getObjectService,
-        IAdRepository adRepository
+        IDeviceRepository deviceRepository
     )
     {
         _logger = logger;
-        _eventCoordinator = eventCoordinator;
         _configRepository = configRepo;
         _backgroundJobClient = backgroundJobClient;
         _deviceRepository = deviceRepository;
-        _mapper = mapper;
-        _presigner = preSignObjectService;
-        _bucketService = getOrCreateBucketService;
-        _putObjectService = putObjectService;
-        _getObjectService = getObjectService;
-        _adRepository = adRepository;
     }
 
     public async Task Run()
@@ -78,16 +53,6 @@ public class ScheduleJob
             pageSize
         );
 
-        var maybeScheduleBucket = await _bucketService.GetSchedule();
-        if (!maybeScheduleBucket.IsSuccessful)
-        {
-            _logger.LogError(
-                "Failed to get bucket info for schedules: {0}",
-                maybeScheduleBucket.Error.ToString()
-            );
-            return;
-        }
-
         var total = await _deviceRepository.Count();
         if (total == 0)
         {
@@ -101,8 +66,7 @@ public class ScheduleJob
 
         foreach (var device in devices)
         {
-            await Calculate(device, maybeScheduleBucket.Value);
-            await _eventCoordinator.SendMessage(device.Id, Command.CallForSchedule);
+            _backgroundJobClient.Enqueue<CalculateJob>(j => j.Calculate(device.Id));
         }
 
         currentPage.Value = pageNumber.ToString();
@@ -138,68 +102,5 @@ public class ScheduleJob
 
             pageNumber = 0;
         }
-    }
-
-    private async Task Calculate(Device device, S3Bucket bucket)
-    {
-        ScheduleContract oldSchedule = new ScheduleContract();
-        var maybeOldSchedule = await _getObjectService.GetObjectByKey(bucket, device.Id.ToString());
-        if (maybeOldSchedule.IsSuccessful)
-        {
-            oldSchedule = maybeOldSchedule.Value;
-        }
-
-        var maybeLastAd = oldSchedule.Schedule.LastOrDefault();
-        var newBatchOfAds = await _adRepository.TakeFrom(
-            maybeLastAd == null ? Guid.Empty : maybeLastAd.Ad!.Id,
-            adPageSize
-        );
-
-        var response = await _putObjectService.Put(
-            bucket,
-            device.Id.ToString(),
-            new MemoryStream(
-                Encoding.UTF8.GetBytes(
-                    JsonConvert.SerializeObject(
-                        new ScheduleContract
-                        {
-                            CreatedAt = DateTime.Now,
-                            DeviceId = device.Id,
-                            Schedule = newBatchOfAds
-                                .Select(async x => new ScheduleItemContract
-                                {
-                                    Ad = _mapper.Map<AdContract>(x),
-                                    DownloadLink = await GetLink(x.Id, bucket)
-                                })
-                                .Select(x => x.Result)
-                        }
-                    )
-                )
-            )
-        );
-        if (!response.IsSuccessful)
-        {
-            _logger.LogError(
-                "Couldn't persist schedule due to error: {0}",
-                response.Error.ToString()
-            );
-        }
-    }
-
-    private async Task<string> GetLink(Guid id, Amazon.S3.Model.S3Bucket bucket)
-    {
-        var maybePresigned = await _presigner.Get(
-            bucket,
-            id.ToString(),
-            DateTime.Now.Add(expiration)
-        );
-        if (!maybePresigned.IsSuccessful)
-        {
-            throw new Exception(
-                string.Format("Received error: {0}", maybePresigned.Error.ToString())
-            );
-        }
-
-        return maybePresigned.Value;
     }
 }
